reservation api created

+++ models/reservation.js
@@ -2,7 +2,7 @@ module.exports = (sequelize, DataTypes)=>{
     const reservation = sequelize.define('reservation',{
         name: DataTypes.STRING,
         email: DataTypes.STRING,
-        phone: DataTypes.INTEGER,
+        phone: DataTypes.BIGINT,
         noOfPerson: DataTypes.STRING,
         time: DataTypes.STRING,
         date: DataTypes.DATE
